Human Resource Management System
In this project we have taken two datasets containing existing and exiting employees.
Both the datasets are explored and cleaned. Then we have developed a model using the Pearson correlation coefficient. We have analysed the model to answer two business questions. What employees are leaving their current jobs? Which type of employees are prone to leave next?
This project can be used by companies to understand the causes of employee resignation and can minimise the resignation rates.